@@ -1,4 +1,4 @@
-import chai from 'chai';
-global.assert = require('assert');
-global.should = require('should');
-global.expect = chai.expect;
+import chai from 'chai'
+global.assert = require('assert')
+global.should = require('should')
+global.expect = chai.expect
