@@ -3,17 +3,17 @@
  */
 
 // Libs
-import { model, Joi } from './lib/model';
-import { adapter } from './lib/adapter';
-import _ from 'lodash';
+import { model, obey } from './lib/model'
+import { adapter } from './lib/adapter'
+import _ from 'lodash'
 
 /**
  * Adds plugins for extending core functionality
  * @param {Function} plugin The plugin function
  */
-const pluginFn = function (plugin) {
-  this[plugin.name] = plugin;
-};
+const pluginFn = function(plugin) {
+  this[plugin.name] = plugin
+}
 
 /**
  * Binds model and adapter to make usable entity
@@ -22,15 +22,15 @@ const pluginFn = function (plugin) {
  */
 const use = (modelName, adapterName) => {
   // Initialize model and adapter
-  const m = model.init(modelName);
-  const a = adapter.init(adapterName);
+  const m = model.init(modelName)
+  const a = adapter.init(adapterName)
   // Return extended (in case model has arbitrary properties/methods)
-  return _.extend(a, m, { plugin: pluginFn });
-};
+  return _.extend(a, m, { plugin: pluginFn })
+}
 
 /**
  * Main modli object
  */
 export {
-  use, model, adapter, Joi
-};
+  use, model, adapter, obey
+}
