@@ -1,7 +1,7 @@
 /* eslint no-unused-expressions: 0 */
 /* global expect, request, describe, it, before, after */
-import '../../setup';
-import { model, Joi } from '../../../src/lib/model';
+import '../../setup'
+import { model } from '../../../src/lib/model'
 
 describe('model', () => {
   // Define models
@@ -10,125 +10,125 @@ describe('model', () => {
     version: 1,
     customProp: 'bar',
     schema: {
-      id: Joi.number().integer(),
-      fname: Joi.string().min(3).max(30),
-      lname: Joi.string().min(3).max(30),
-      roles: {
-        admin: Joi.boolean()
-      }
+      id: { type: 'number' },
+      fname: { type: 'string' },
+      lname: { type: 'string' },
+      roles: { type: 'object', keys: {
+        admin: { type: 'boolean', default: false }
+      }}
     }
-  };
+  }
 
   const modelObjV2 = {
     name: 'foo',
     version: 2,
     customProp: 'baz',
     schema: {
-      id: Joi.number().integer(),
-      fname: Joi.string().min(3).max(30),
-      lname: Joi.string().min(3).max(30),
-      roles: {
-        admin: Joi.boolean()
-      },
-      email: Joi.string().email().min(3).max(30).required()
+      id: { type: 'number' },
+      fname: { type: 'string' },
+      lname: { type: 'string' },
+      roles: { type: 'object', keys: {
+        admin: { type: 'boolean', default: false }
+      }},
+      email: { type: 'email', required: 'true' }
     }
-  };
+  }
 
   // Hoist placeholder for sharing across tests
-  let testModel = {};
+  let testModel = {}
 
   describe('add', () => {
     it('fails if missing properties in the model obejct', () => {
       try {
-        model.add({});
+        model.add({})
       } catch (e) {
-        expect(e).to.be.an.instanceof(Error);
+        expect(e).to.be.an.instanceof(Error)
       }
-    });
+    })
 
     it('adds a new model entry to the model object store', () => {
       // Add the first version
-      model.add(modelObjV1);
+      model.add(modelObjV1)
       // Add the second version
-      model.add(modelObjV2);
+      model.add(modelObjV2)
       // Ensure entry (against V2)
-      expect(model.store).to.have.property(modelObjV2.name);
+      expect(model.store).to.have.property(modelObjV2.name)
       // Ensure version (against V2)
-      expect(model.store[modelObjV2.name]).to.have.property(modelObjV2.version);
-    });
-  });
+      expect(model.store[modelObjV2.name]).to.have.property(modelObjV2.version)
+    })
+  })
 
   describe('init', () => {
     // Invalid model check
     it('fails if invalid model defined', () => {
       try {
-        model.init(null);
+        model.init(null)
       } catch (e) {
-        expect(e).to.be.an.instanceof(Error);
+        expect(e).to.be.an.instanceof(Error)
       }
-    });
+    })
     // Initializes model
     it('initializes the model by binding validation and versioning', () => {
-      testModel = model.init('foo');
-      expect(testModel).to.be.an.object;
-    });
-  });
+      testModel = model.init('foo')
+      expect(testModel).to.be.an.object
+    })
+  })
 
   // Define pass data
   const testPassDataV1 = {
     id: 12345,
     fname: 'John',
     lname: 'Doe'
-  };
+  }
 
   const testPassDataV2 = {
     id: 12345,
     fname: 'John',
     lname: 'Doe',
     email: 'jdoe@gmail.com'
-  };
+  }
 
   // Define fail data
-  const testFailDataV1 = testPassDataV2;
+  const testFailDataV1 = testPassDataV2
 
   const testFailDataV2 = {
     id: 'foo',
     fname: 123,
     lname: [ 'bar' ],
     email: 'jdoe[at]gmail.com'
-  };
+  }
 
   describe('validate', () => {
     // Pass validation condition
     it('passes validation when object matches rules', () => {
-      const passTest = testModel.validate(testPassDataV2);
-      expect(passTest).to.be.null;
-    });
+      const passTest = testModel.validate(testPassDataV2)
+      expect(passTest).to.be.null
+    })
     // Fail validation condition
     it('fails validation when object does not match rules', () => {
-      const failTest = testModel.validate(testFailDataV2);
-      expect(failTest).to.not.be.null;
-    });
+      const failTest = testModel.validate(testFailDataV2)
+      expect(failTest).to.not.be.null
+    })
     // Use custom formatter
     it('uses a custom validation error format when specified', () => {
       // Define custom formatter
       model.customValidationError = (err) => {
-        return err.details[0].message;
-      };
-      const testCustom = testModel.validate(testFailDataV2);
-      expect(testCustom).to.equal('"id" must be a number');
-    });
+        return err.details[0].message
+      }
+      const testCustom = testModel.validate(testFailDataV2)
+      expect(testCustom).to.equal('"id" must be a number')
+    })
     // Pass on older version
     it('passes validation on older version of schema', () => {
-      const passTest = testModel.validate(testPassDataV1, 1);
-      expect(passTest).to.be.null;
-    });
+      const passTest = testModel.validate(testPassDataV1, 1)
+      expect(passTest).to.be.null
+    })
     // Fail on older version
     it('fails validation on older version of schema', () => {
-      const failTest = testModel.validate(testFailDataV1, 1);
-      expect(failTest).to.not.be.null;
-    });
-  });
+      const failTest = testModel.validate(testFailDataV1, 1)
+      expect(failTest).to.not.be.null
+    })
+  })
 
   describe('sanitize', () => {
     it('removes items from the data which are not present in the schema', () => {
@@ -139,14 +139,14 @@ describe('model', () => {
           admin: true
         },
         email: 'jdoe@gmail.com' // <- not in v.1 of the model
-      }, 1);
+      }, 1)
       expect(sanitized).to.deep.equal({
         fname: 'John',
         lname: 'Smith',
         roles: {
           admin: true
         }
-      });
-    });
-  });
-});
+      })
+    })
+  })
+})
